airtable_sync/requirements.txt
psycopg2-binary==2.9.10
pyairtable==2.3.3
